save point 1

+++ src/main/java/com/service/ICrudService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface ICrudService<E> {
     List<E> getAll();
-    E getById(int id);
+    E getById(long id);
     E create(E e);
     E edit (E e);
     void deleteById(int id);
