+++ src/main/java/com/service/ipml/ReportServiceImpl.java
@@ -18,7 +18,7 @@ public class ReportServiceImpl implements IReportService {
     }
 
     @Override
-    public Report getById(int id) {
+    public Report getById(long id) {
         return null;
     }
